+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/RealMiniApp.kt
@@ -11,7 +11,7 @@ import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermission
 import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionCache
 import com.rakuten.tech.mobile.miniapp.storage.CachedManifest
 import com.rakuten.tech.mobile.miniapp.storage.DownloadedManifestCache
-import com.rakuten.tech.mobile.miniapp.storage.MiniAppManifestVerifier
+import com.rakuten.tech.mobile.miniapp.storage.verifier.MiniAppManifestVerifier
 
 @Suppress("TooManyFunctions", "LongMethod")
 internal class RealMiniApp(
