+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/storage/verifier/CachedMiniAppVerifier.kt
@@ -1,45 +1,36 @@
-package com.rakuten.tech.mobile.miniapp.storage
+package com.rakuten.tech.mobile.miniapp.storage.verifier
 
 import android.content.Context
 import android.content.SharedPreferences
 import android.util.Base64
 import android.util.Log
 import androidx.annotation.VisibleForTesting
-import androidx.security.crypto.EncryptedSharedPreferences
-import androidx.security.crypto.MasterKeys
-import com.rakuten.tech.mobile.miniapp.MiniAppVerificationException
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.async
-import kotlinx.coroutines.withContext
 import java.io.File
 import java.security.MessageDigest
 
-internal class CachedMiniAppVerifier
-@VisibleForTesting constructor(
+internal class CachedMiniAppVerifier @VisibleForTesting constructor(
     private val prefs: SharedPreferences,
     private val coroutineDispatcher: CoroutineDispatcher
 ) {
+    private val storeHashVerifier = StoreHashVerifier(prefs, coroutineDispatcher)
+
     constructor(context: Context) : this(
-        prefs = initEncryptedSharedPreference(context),
+        prefs = initEncryptedSharedPreference(
+            context,
+            "com.rakuten.tech.mobile.miniapp.cache.hash"
+        ),
         coroutineDispatcher = Dispatchers.IO
     )
 
     /** Verifies that the cached files for the Mini App have not been modified. */
-    fun verify(appId: String, directory: File): Boolean {
-        val hash = calculateHash(directory)
-        val storedHash = prefs.getString(appId, null) ?: ""
-
-        return hash == storedHash
-    }
+    fun verify(appId: String, directory: File): Boolean =
+        storeHashVerifier.verify(appId, calculateHash(directory))
 
     /** Stores hash in encrypted shared preferences. This runs asynchronously so it will return immediately. */
-    suspend fun storeHashAsync(appId: String, directory: File) = withContext(coroutineDispatcher) {
-        async {
-            val hash = calculateHash(directory)
-            prefs.edit().putString(appId, hash).apply()
-        }
-    }
+    suspend fun storeHashAsync(appId: String, directory: File) =
+        storeHashVerifier.storeHashAsync(appId, calculateHash(directory))
 
     @SuppressWarnings("TooGenericExceptionCaught", "LongMethod", "NestedBlockDepth")
     private fun calculateHash(
@@ -70,17 +61,4 @@ internal class CachedMiniAppVerifier
         private val TAG = this::class.simpleName
         private const val BUFFER_SIZE = 8192
     }
-}
-
-@Suppress("TooGenericExceptionCaught", "SwallowedException")
-private fun initEncryptedSharedPreference(context: Context) = try {
-    EncryptedSharedPreferences.create(
-        "com.rakuten.tech.mobile.miniapp.cache.hash",
-        MasterKeys.getOrCreate(MasterKeys.AES256_GCM_SPEC),
-        context,
-        EncryptedSharedPreferences.PrefKeyEncryptionScheme.AES256_SIV,
-        EncryptedSharedPreferences.PrefValueEncryptionScheme.AES256_GCM
-    )
-} catch (e: Exception) {
-    throw MiniAppVerificationException(e.message)
 }
