+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/storage/verifier/StoreHashVerifier.kt
@@ -1,0 +1,43 @@
+package com.rakuten.tech.mobile.miniapp.storage.verifier
+
+import android.content.Context
+import android.content.SharedPreferences
+import androidx.annotation.VisibleForTesting
+import androidx.security.crypto.EncryptedSharedPreferences
+import androidx.security.crypto.MasterKeys
+import com.rakuten.tech.mobile.miniapp.MiniAppVerificationException
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.async
+import kotlinx.coroutines.withContext
+
+internal class StoreHashVerifier @VisibleForTesting constructor(
+    private val prefs: SharedPreferences,
+    private val coroutineDispatcher: CoroutineDispatcher
+) {
+    /** Verifies that the cached data has not been modified. */
+    fun verify(appId: String, hash: String): Boolean {
+        val storedHash = prefs.getString(appId, null) ?: ""
+        return hash == storedHash
+    }
+
+    /** Stores hash in encrypted shared preferences. This runs asynchronously so it will return immediately. */
+    suspend fun storeHashAsync(appId: String, hash: String) =
+        withContext(coroutineDispatcher) {
+            async {
+                prefs.edit().putString(appId, hash).apply()
+            }
+        }
+}
+
+@Suppress("TooGenericExceptionCaught", "SwallowedException")
+internal fun initEncryptedSharedPreference(context: Context, fileName: String) = try {
+    EncryptedSharedPreferences.create(
+        fileName,
+        MasterKeys.getOrCreate(MasterKeys.AES256_GCM_SPEC),
+        context,
+        EncryptedSharedPreferences.PrefKeyEncryptionScheme.AES256_SIV,
+        EncryptedSharedPreferences.PrefValueEncryptionScheme.AES256_GCM
+    )
+} catch (e: Exception) {
+    throw MiniAppVerificationException(e.message)
+}
